requirements.txt
aiofiles==24.1.0
aiosmtplib==4.0.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
firebase-admin==6.5.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
passlib==1.7.4
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
